accept/reject lien ok

+++ src/Controller/Api/ApiHandleLinksController.php
@@ -5,9 +5,11 @@ declare(strict_types=1);
 namespace Src\Controller\Api;
 
 
-class ApiHandleLinksController extends ApiController{
+class ApiHandleLinksController extends ApiController
+{
 
-    public function sendFriendRequest(){
+    public function sendFriendRequest()
+    {
 
         try {
             if ($_SERVER['REQUEST_METHOD'] !== 'POST') {
@@ -48,9 +50,9 @@ class ApiHandleLinksController extends ApiController{
             error_log($th->getMessage());
             $this->sendJson(["message" => "Erreur serveur"], 500);
         }
-
     }
-    public function cancelRequest(){
+    public function cancelRequest()
+    {
 
         try {
             if ($_SERVER['REQUEST_METHOD'] !== 'POST') {
@@ -91,9 +93,9 @@ class ApiHandleLinksController extends ApiController{
             error_log($th->getMessage());
             $this->sendJson(["message" => "Erreur serveur"], 500);
         }
-
     }
-    public function breakLink(){
+    public function breakLink()
+    {
 
         try {
             if ($_SERVER['REQUEST_METHOD'] !== 'POST') {
@@ -117,11 +119,11 @@ class ApiHandleLinksController extends ApiController{
             }
 
             // //  vérifier si le lien existe déjà
-            // $linkExists = $this->apiHandleLinksModel->checkIfLinkExists($userId, $friendId);
-            // if (!$linkExists) {
-            //     $this->sendJson(["message" => "Aucun lien en cours"], 409);
-            //     return;
-            // }
+            $linkExists = $this->apiHandleLinksModel->checkIfLinkExists($userId, $friendId);
+            if (!$linkExists) {
+                $this->sendJson(["message" => "Aucun lien en cours"], 409);
+                return;
+            }
 
             $result = $this->apiHandleLinksModel->deleteLink($userId, $friendId);
 
@@ -134,6 +136,91 @@ class ApiHandleLinksController extends ApiController{
             error_log($th->getMessage());
             $this->sendJson(["message" => "Erreur serveur"], 500);
         }
+    }
+    public function acceptRequest()
+    {
 
+        try {
+            if ($_SERVER['REQUEST_METHOD'] !== 'POST') {
+                $this->sendJson(["message" => "Méthode non autorisée"], 405);
+                return;
+            }
+
+            $userId = $this->securityApiController->getAuthenticatedUserIdFromToken();
+
+            if (!$userId) {
+                $this->sendJson(["message" => "Utilisateur non authentifié"], 401);
+                return;
+            }
+
+            $data = json_decode(file_get_contents('php://input'), true);
+            $friendId = $data['id'] ?? null;
+
+            if (!$friendId) {
+                $this->sendJson(["message" => "ID d'ami manquant"], 400);
+                return;
+            }
+
+            //  vérifier si le lien existe déjà
+            $linkExists = $this->apiHandleLinksModel->checkIfLinkExists($userId, $friendId);
+            if (!$linkExists) {
+                $this->sendJson(["message" => "Aucune demande en cours"], 409);
+                return;
+            }
+
+            $result = $this->apiHandleLinksModel->acceptFriendRequest($userId, $friendId);
+
+            if ($result) {
+                $this->sendJson(["message" => "Demande d'ami acceptée avec succès"]);
+            } else {
+                $this->sendJson(["message" => "Échec de l'acceptation de la demande d'ami"], 500);
+            }
+        } catch (\Throwable $th) {
+            error_log($th->getMessage());
+            $this->sendJson(["message" => "Erreur serveur"], 500);
+        }
+    }
+    public function declineRequest()
+    {
+
+        try {
+            if ($_SERVER['REQUEST_METHOD'] !== 'POST') {
+                $this->sendJson(["message" => "Méthode non autorisée"], 405);
+                return;
+            }
+
+            $userId = $this->securityApiController->getAuthenticatedUserIdFromToken();
+
+            if (!$userId) {
+                $this->sendJson(["message" => "Utilisateur non authentifié"], 401);
+                return;
+            }
+
+            $data = json_decode(file_get_contents('php://input'), true);
+            $friendId = $data['id'] ?? null;
+
+            if (!$friendId) {
+                $this->sendJson(["message" => "ID d'ami manquant"], 400);
+                return;
+            }
+
+            //  vérifier si le lien existe déjà
+            $linkExists = $this->apiHandleLinksModel->checkIfLinkExists($userId, $friendId);
+            if (!$linkExists) {
+                $this->sendJson(["message" => "Aucune demande en cours"], 409);
+                return;
+            }
+
+            $result = $this->apiHandleLinksModel->rejectFriendRequest($userId, $friendId);
+
+            if ($result) {
+                $this->sendJson(["message" => "Demande d'ami rejetée avec succès"]);
+            } else {
+                $this->sendJson(["message" => "Échec du rejet de la demande d'ami"], 500);
+            }
+        } catch (\Throwable $th) {
+            error_log($th->getMessage());
+            $this->sendJson(["message" => "Erreur serveur"], 500);
+        }
     }
 }
